Added Grave Key Texture
The Grave Key has a texture now when using the ressourcepack. If not, the Key will now look like a poisonous potato.

The Potato has been choosen because with the component consumable removed, its truely a completely useless item. which means that the key has no hidden second functionality.

The retexturing uses the new custom model data introduced in 1.21.4.
This means that the datapack is no longer compatible with older Versions!