Added support for all 1.21.x Versions again
In the past, the datapack was compatible with older versions.

But by adding a texture to the key, the methods differentiated into different versions. That resulted in two releases for different versions.

But by using the overlay method, it is now possible to have multiple versions of the same datapack in one pack.

That makes the datapack compatible with all 1.21.x versions again.